# BE/docs/swagger.yaml
swagger: "2.0"
info:
  title: IT4409 Jira-Clone API
  description: |
    Backend API cho hệ thống quản lý dự án kiểu Jira.
    Clean Architecture (Go) + PostgreSQL (Supabase).
    
    **Hướng dẫn xác thực:**
    1. Gọi `/api/auth/register` hoặc `/api/auth/login` để lấy JWT token.
    2. Nhấn nút "Authorize" ở góc trên, nhập: `Bearer <token>`.
    3. Tất cả API có khóa 🔒 đều yêu cầu JWT.
  version: "2.0"
  contact:
    name: IT4409 Team
basePath: /
schemes:
  - http
securityDefinitions:
  BearerAuth:
    description: "Nhập JWT token. Ví dụ: Bearer eyJhbGciOiJI..."
    type: apiKey
    name: Authorization
    in: header

# ════════════════════════════════════════════════
#  DEFINITIONS (Cấu trúc dữ liệu)
# ════════════════════════════════════════════════
definitions:
...
  Project:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      key:
        type: string
      description:
        type: string
      type:
        type: string
        enum: [kanban, scrum]
      lead_id:
        type: string
      icon:
        type: string
      created_by:
        type: string
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  ProjectMember:
    type: object
    properties:
      project_id:
        type: string
      user_id:
        type: string
      role:
        type: string
        enum: [admin, member, viewer]
      joined_at:
        type: string
        format: date-time
...
  Label:
    type: object
    properties:
      id:
        type: string
      projectId:
        type: string
      name:
        type: string
      color:
        type: string
      createdAt:
        type: string
        format: date-time
...
paths:
...
  /api/projects:
    post:
      tags: [Projects]
      summary: Tạo dự án mới
      description: Tạo dự án + khởi tạo counter + thêm người tạo làm admin.
      security: [{BearerAuth: []}]
      consumes: [application/json]
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name, key]
            properties:
              name:
                type: string
                example: Dự án Kanban
              key:
                type: string
                example: KANBAN
              description:
                type: string
                example: Quản lý công việc theo phương pháp Kanban
              type:
                type: string
                enum: [kanban, scrum]
                default: kanban
      responses:
        201:
          description: Tạo dự án thành công
          schema:
            $ref: "#/definitions/Project"
        401:
          description: Chưa đăng nhập
    get:
      tags: [Projects]
      summary: Liệt kê dự án của tôi
      security: [{BearerAuth: []}]
      responses:
        200:
          description: Danh sách dự án mà người dùng là thành viên
          schema:
            type: array
            items:
              $ref: "#/definitions/Project"
  /api/projects/{projectID}:
    get:
      tags: [Projects]
      summary: Chi tiết dự án
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
      responses:
        200:
          description: Thông tin dự án
          schema:
            $ref: "#/definitions/Project"
        403:
          description: Không có quyền
        404:
          description: Không tìm thấy
    patch:
      tags: [Projects]
      summary: Cập nhật dự án
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              icon:
                type: string
      responses:
        200:
          description: Cập nhật thành công
        403:
          description: Không có quyền admin
    delete:
      tags: [Projects]
      summary: Xóa dự án (xóa mềm)
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
      responses:
        204:
          description: Xóa thành công
        403:
          description: Không có quyền admin

  # ── PROJECT MEMBERS (Người B) ──────────────────
  /api/projects/{projectID}/members:
    get:
      tags: [Project Members]
      summary: Liệt kê thành viên
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
      responses:
        200:
          description: Danh sách thành viên
          schema:
            type: array
            items:
              $ref: "#/definitions/ProjectMember"
    post:
      tags: [Project Members]
      summary: Thêm thành viên
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              userId:
                type: string
              role:
                type: string
                enum: [admin, member, viewer]
      responses:
        201:
          description: Thêm thành viên thành công
        403:
          description: Không có quyền admin
  /api/projects/{projectID}/members/{userID}:
    put:
      tags: [Project Members]
      summary: Đổi vai trò thành viên
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
        - in: path
          name: userID
          type: string
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              role:
                type: string
                enum: [admin, member, viewer]
      responses:
        200:
          description: Đổi vai trò thành công
    delete:
      tags: [Project Members]
      summary: Xóa thành viên khỏi dự án
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
        - in: path
          name: userID
          type: string
          required: true
      responses:
        204:
          description: Xóa thành viên thành công

  # ── BOARDS (Người B) ───────────────────────────
...
  /api/projects/{projectID}/labels:
    post:
      tags: [Labels]
      summary: Tạo nhãn cho dự án
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name]
            properties:
              name:
                type: string
                example: Bug
              color:
                type: string
                example: "#ef4444"
                default: "#6366f1"
      responses:
        201:
          description: Tạo nhãn thành công
          schema:
            $ref: "#/definitions/Label"
    get:
      tags: [Labels]
      summary: Liệt kê nhãn của dự án
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
      responses:
        200:
          description: Danh sách nhãn
          schema:
            type: array
            items:
              $ref: "#/definitions/Label"
  /api/labels/{labelID}:
    patch:
      tags: [Labels]
      summary: Sửa nhãn
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: labelID
          type: string
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              name:
                type: string
              color:
                type: string
      responses:
        200:
          description: Sửa nhãn thành công
    delete:
      tags: [Labels]
      summary: Xóa nhãn
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: labelID
          type: string
          required: true
      responses:
        204:
          description: Xóa nhãn thành công
  /api/issues/{issueKey}/labels:
    post:
      tags: [Labels]
      summary: Gắn nhãn vào công việc
      description: "Yêu cầu quyền: member+"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: issueKey
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              labelId:
                type: string
      responses:
        201:
          description: Gắn nhãn thành công
  /api/issues/{issueKey}/labels/{labelID}:
    delete:
      tags: [Labels]
      summary: Gỡ nhãn khỏi công việc
      description: "Yêu cầu quyền: member+"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: issueKey
          type: string
          required: true
        - in: path
          name: labelID
          type: string
          required: true
      responses:
        204:
          description: Gỡ nhãn thành công

  # ── ISSUES ──────────────────────────────────────
...